# api-stock/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.AppError:
    properties:
      code:
        description: |-
          HTTP status code
          Required: true
          Example: 400
        type: integer
      message:
        description: |-
          Descriptive error message
          Required: true
          Example: "Invalid request parameters"
        type: string
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@stockapi.com
    name: API Support
  description: API for stock recommendations with machine learning features
  license:
    name: MIT
  title: Stock Recommendation API
  version: "1.0"
paths:
...
  /http/v1/recommendations:
    get:
      consumes:
      - application/json
      description: Get paginated list of stock recommendations, filterable by ticker
      parameters:
      - description: Stock ticker to filter by
        in: query
        name: ticker
        type: string
      - default: 1
        description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns recommendations and pagination info
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Get stock recommendations
      tags:
      - recommendations
  /http/v1/recommendations/best:
    get:
      consumes:
      - application/json
      description: Get top stock recommendations based on scoring model
      parameters:
      - default: 5
        description: Number of recommendations to return
        in: query
        maximum: 20
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns best recommendations and generation timestamp
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Get best stock recommendations
      tags:
      - recommendations
  /http/v1/recommendations/tickers:
    get:
      consumes:
      - application/json
      description: Get list of all available stock tickers with recommendations
      produces:
      - application/json
      responses:
        "200":
          description: List of tickers
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Get available stock tickers
      tags:
      - recommendations
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"